Transformado toda a aplicação

+++ src/modules/base/service/base.service.ts
@@ -1,26 +1,29 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  HttpStatus,
+  Inject,
+  Injectable,
+  forwardRef,
+} from '@nestjs/common';
 import { LoggerServer } from 'src/loggerServer';
 
-import { BaseDTO } from 'src/DTO/base.dto';
+import { BaseDTO } from '../interfaces/base.dto';
 
-import * as csvConverter from 'csvjson-csv2json';
-import { Base } from 'src/Mongo/Interface/base.interface';
-import { FilterDTO } from 'src/Mongo/Interface/query.interface';
-import {
-  StatesInterface,
-  StructuresInterface,
-} from 'src/Mongo/Interface/structures.interface';
-import { BaseRepository } from 'src/Mongo/repository/base.repository';
-import {
-  DengueStructure,
-  TesteStructure,
-} from 'src/modules/base/structures.object';
+import { parse } from 'papaparse';
+import { FilterDTO } from 'src/interfaces/query.interface';
+import { ParametersService } from 'src/modules/parameters/services/parameters.service';
+import { StructureService } from 'src/modules/structure/service/structure.service';
+import { BaseRepository } from '../base.repository';
+import { Base } from '../interfaces/base.interface';
 
 @Injectable()
 export class BaseService {
   constructor(
     private readonly baseRepository: BaseRepository,
     private readonly logger: LoggerServer,
+    private readonly parametersService: ParametersService,
+    @Inject(forwardRef(() => StructureService))
+    private readonly structureService: StructureService,
   ) {}
 
   convertJSON(file: Express.Multer.File) {
@@ -30,24 +33,25 @@ export class BaseService {
         `Arquivo grande, pode demorar mais que o normal. Tamanho: ${file.size} bytes aproximadamente | Nome: ${file.originalname}`,
       );
 
-    return csvConverter(file.buffer.toString('utf-8'), {
-      parseNumbers: true,
-    });
+    return parse(file.buffer.toString('utf-8'), {
+      delimiter: ';',
+      dynamicTyping: true,
+      header: true,
+    }).data;
   }
 
   async uploadFiles(
     files: Array<Express.Multer.File>,
-    structure: string,
+    structureID: string,
     name: string,
+    userID: string,
   ) {
     if (files === undefined || files.length === 0) {
       this.logger.error('Enviado arquivos inválidos ou nenhum arquivo enviado');
       throw new HttpException('No files uploaded', HttpStatus.BAD_REQUEST);
     }
 
-    structure = structure.toLowerCase();
-
-    const structureObject = this.getStructureByName(structure);
+    const structureObject = await this.structureService.getByID(structureID);
     if (!structureObject)
       throw new HttpException(
         'Estrutura não encontrada',
@@ -64,31 +68,44 @@ export class BaseService {
       `Fazendo o upload dos arquivos e convertendo para JSON. Número de arquivos: ${files.length}`,
     );
 
-    //Isso permite com que o sistema possa adicionar novos tipos de simulação
-    const structureFinal = new BaseDTO();
+    const structureFinal = {} as Base;
     structureFinal.name = name;
-    structureFinal.parameters = {} as any;
-    structureFinal.type = structure as any;
-    const params = Object.keys(structureObject.type_parameters);
+    structureFinal.type = structureID;
+    structureFinal.user = userID;
+    structureFinal.parameters = {};
 
-    for (let i = 0; i < params.length; i++) {
-      const param = params[i];
-      const subParams = Object.keys(structureObject.type_parameters[param]);
+    const parameters = structureObject.parameters;
 
-      if (subParams.length === 0)
-        structureFinal.parameters[param] = this.convertJSON(files[i]);
-      else {
-        for (const subParam of subParams) {
-          structureFinal.parameters[param][subParam] = this.convertJSON(
-            files[i],
-          );
-          i++;
+    try {
+      for (let i = 0; i < parameters.length; i++) {
+        const param = parameters[i];
+
+        if (param.subParameters.length === 0)
+          structureFinal.parameters[param.name] = this.convertJSON(files[i]);
+        else {
+          for (const subParam of param.subParameters) {
+            structureFinal.parameters[param.name][subParam.name] =
+              this.convertJSON(files[i]);
+
+            i++;
+          }
         }
       }
+    } catch (e) {
+      throw new HttpException(
+        'Erro ao converter os arquivos! Valide se os arquivos estão no formato correto',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
     }
 
     try {
-      return await this.baseRepository.saveBase(structureFinal);
+      const parameters = await this.parametersService.uploadAllParameters(
+        structureFinal.parameters,
+      );
+      return await this.baseRepository.saveBase({
+        ...structureFinal,
+        parameters,
+      });
     } catch (e) {
       throw new HttpException(
         `Erro ao salvar a base ${e}`,
@@ -97,55 +114,8 @@ export class BaseService {
     }
   }
 
-  async getBases(query: FilterDTO): Promise<Base[]> {
+  async getBases(query: FilterDTO) {
     return await this.baseRepository.getBases(query);
-  }
-
-  async getParameters(parametersID: string): Promise<object> {
-    return await this.baseRepository.readFile(parametersID);
-  }
-
-  getAllStructures(): StructuresInterface[] {
-    return [DengueStructure, TesteStructure];
-  }
-
-  getStructureByName(structureName: string): StructuresInterface {
-    const structures = this.getAllStructures();
-    const structure = structures.find(
-      (elemm) => elemm.name.toLowerCase() === structureName.toLowerCase(),
-    );
-    if (!structure) {
-      throw new HttpException('Estrutura não encontrada', HttpStatus.NOT_FOUND);
-    } else return structure;
-  }
-
-  async getStructureByID(baseID: string): Promise<StructuresInterface> {
-    const base = await this.baseRepository.getBaseByID(baseID);
-    if (!base)
-      throw new HttpException(
-        'Nenhuma base encontrada com esse ID',
-        HttpStatus.NOT_FOUND,
-      );
-    const structure = this.getAllStructures().find(
-      (elem) => elem.name.toLowerCase() === base.type.toLowerCase(),
-    );
-    if (structure === undefined)
-      throw new HttpException('Estrutura não encontrada', HttpStatus.NOT_FOUND);
-    else return structure;
-  }
-
-  getStatesByStructure(name: string): StatesInterface {
-    const structure = this.getStructureByName(name);
-    if (!structure) return [];
-    return structure.states;
-  }
-
-  async getStatesByBase(baseID: string): Promise<StatesInterface> {
-    const structure = await this.getStructureByID(baseID);
-    if (!structure) {
-      throw new HttpException('Estrutura não encontrada', HttpStatus.NOT_FOUND);
-    }
-    return this.getStatesByStructure(structure.name);
   }
 
   async getBaseByID(baseID: string): Promise<Base> {
@@ -160,23 +130,68 @@ export class BaseService {
     return base;
   }
 
-  async saveBase(newBase: BaseDTO): Promise<Base> {
-    return await this.baseRepository.saveBase(newBase);
+  async saveBase(
+    base: BaseDTO,
+    structureID: string,
+    userID: string,
+  ): Promise<Base> {
+    const structure = await this.structureService.getByID(structureID);
+    if (!structure)
+      throw new HttpException(
+        'Estrutura não encontrada',
+        HttpStatus.BAD_REQUEST,
+      );
+
+    try {
+      const parameters =
+        await this.structureService.createParametersObject(structureID);
+
+      const parametersObject =
+        await this.parametersService.uploadAllParameters(parameters);
+
+      return await this.baseRepository.saveBase({
+        ...base,
+        user: userID,
+        type: structureID,
+        parameters: parametersObject,
+      });
+    } catch (e) {
+      this.logger.error(`Erro ao salvar a base! Erro: ${e}`);
+
+      throw new HttpException(
+        'Erro ao salvar base',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
   }
 
-  async updateBase(baseID: string, newBase: BaseDTO): Promise<Base> {
-    try {
-      const base = await this.baseRepository.getBaseByID(baseID);
-      if (!base)
-        throw new HttpException(
-          'Nenhuma base encontrada com esse ID',
-          HttpStatus.NOT_FOUND,
-        );
+  async updateBase(
+    baseID: string,
+    newBase: BaseDTO,
+    userID: string,
+  ): Promise<Base> {
+    const base = await this.baseRepository.getBaseByID(baseID);
+    if (!base)
+      throw new HttpException(
+        'Nenhuma base encontrada com esse ID',
+        HttpStatus.NOT_FOUND,
+      );
 
-      await this.baseRepository.updateBase(baseID, newBase, base.parametersID);
+    if (base.user?.toString() !== userID)
+      throw new HttpException(
+        'Você não tem permissão para editar essa base',
+        HttpStatus.UNAUTHORIZED,
+      );
+
+    try {
+      const updatedBase = await this.baseRepository.updateBase(baseID, {
+        ...base,
+        ...newBase,
+        updatedAt: new Date().toISOString(),
+      });
 
       this.logger.warn(`Base atualizada: ${newBase.name}`);
-      return this.baseRepository.getBaseByID(baseID);
+      return updatedBase;
     } catch (e) {
       this.logger.error(`Erro ao atualizar a base! Erro: ${e}`);
 
@@ -187,13 +202,28 @@ export class BaseService {
     }
   }
 
-  async deleteBase(baseID: string): Promise<Base> {
+  async deleteBase(baseID: string, userID: string): Promise<Base> {
+    const base = await this.baseRepository.getBaseByID(baseID);
+
+    if (!base)
+      throw new HttpException(
+        'Nenhuma base encontrada com esse ID',
+        HttpStatus.NOT_FOUND,
+      );
+
+    if (base.user.toString() !== userID)
+      throw new HttpException(
+        'Você não tem permissão para deletar essa base',
+        HttpStatus.UNAUTHORIZED,
+      );
+
     try {
       const baseDeleted = await this.baseRepository.deleteBase(baseID);
 
       this.logger.warn(`Base deletada: ${baseDeleted.name}`);
       return baseDeleted;
     } catch (e) {
+      this.logger.error(`Erro ao deletar a base! Erro: ${e}`);
       throw new HttpException(
         'Algum erro ocorreu ao deletar a base',
         HttpStatus.NOT_FOUND,
